quit/requirements.txt
accelerate==1.8.1
addict==2.4.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
async-timeout==5.0.1
attrs==25.3.0
av==14.4.0
bitsandbytes==0.41.0
blis==0.7.11
catalogue==2.0.10
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
cloudpathlib==0.21.1
confection==0.1.5
contourpy==1.3.2
cycler==0.12.1
cymem==2.0.11
datasets==2.16.1
decord==0.6.0
deepspeed==0.14.4
defusedxml==0.7.1
dill==0.3.7
docstring_parser==0.16
einops==0.6.1
einops-exts==0.0.4
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
exceptiongroup==1.3.0
fastapi==0.115.14
filelock==3.18.0
flash-attn @ file:///root/flash_attn-2.7.0.post2%2Bcu11torch2.1cxx11abiFALSE-cp310-cp310-linux_x86_64.whl#sha256=e0ff1d33d44a2d9b93b672a92a1663329439d14e259fa60068589d1d7202b078
fonttools==4.58.4
frozenlist==1.7.0
fsspec==2023.10.0
ftfy==6.3.1
gitdb==4.0.12
GitPython==3.1.44
gradio_client==0.2.9
h11==0.14.0
hf-xet==1.1.5
hf_transfer==0.1.9
hjson==3.1.0
httpcore==0.17.3
httpx==0.24.0
huggingface-hub==0.33.1
idna==3.10
imageio==2.37.0
Jinja2==3.1.6
joblib==1.5.1
kiwisolver==1.4.8
kornia==0.8.1
kornia_rs==0.1.9
langcodes==3.5.0
language_data==1.3.0
latex2mathml==3.78.0
lazy_loader==0.4
# Editable install with no version control (llava==1.7.0.dev0)
-e /root/autodl-tmp/code/ZoomEye
marisa-trie==1.2.1
markdown-it-py==3.0.0
markdown2==2.5.3
MarkupSafe==3.0.2
matplotlib==3.10.3
mdurl==0.1.2
mpmath==1.3.0
multidict==6.6.3
multiprocess==0.70.15
murmurhash==1.0.13
networkx==3.4.2
ninja==1.11.1.4
numpy==1.26.1
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-ml-py==12.575.51
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.9.86
nvidia-nvtx-cu12==12.1.105
open_clip_torch==2.32.0
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
packaging==25.0
pandas==2.3.0
patsy==1.0.1
peft==0.4.0
pillow==11.2.1
platformdirs==4.3.8
preshed==3.0.10
progressbar2==4.5.0
propcache==0.3.2
protobuf==6.32.1
psutil==7.0.0
py-cpuinfo==9.0.0
pyarrow==20.0.0
pyarrow-hotfix==0.7
pycocotools==2.0.10
pydantic==1.10.8
pydantic_core==2.33.2
pygam==0.9.1
Pygments==2.19.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-utils==3.9.1
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.4
rich==14.0.0
safetensors==0.5.3
scikit-image==0.25.2
scikit-learn==1.2.2
scipy==1.11.4
seaborn==0.13.2
sentencepiece==0.2.1
sentry-sdk==2.32.0
setproctitle==1.3.6
shellingham==1.5.4
shortuuid==1.0.13
shtab==1.7.2
six==1.17.0
smart_open==7.3.0.post1
smmap==5.0.2
sniffio==1.3.1
spacy==3.7.6
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.1
starlette==0.46.2
statsmodels==0.14.5
supervision==0.3.2
svgwrite==1.4.3
sympy==1.14.0
thinc==8.2.5
threadpoolctl==3.6.0
tifffile==2025.5.10
timm==1.0.16
tokenizers==0.20.3
tomli==2.2.1
torch==2.1.2+cu118
torchvision==0.16.2+cu118
tqdm==4.67.1
transformers==4.46.0
triton==2.1.0
typeguard==4.4.4
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.0
tyro==0.9.25
tzdata==2025.2
urllib3==1.26.20
uvicorn==0.35.0
wandb==0.20.1
wasabi==1.1.3
wavedrom==2.0.3.post3
wcwidth==0.2.13
weasel==0.4.1
websockets==15.0.1
wrapt==1.17.2
xxhash==3.5.0
yapf==0.43.0
yarl==1.20.1